services/llm-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
confluent-kafka==2.5.0
sentence-transformers==3.0.1
requests==2.32.3
psycopg2-binary==2.9.9
redis==5.0.8
tenacity==9.0.0
python-dotenv==1.0.1
